TrinoClusters must specify a catalogLabelSelector (#277)
# Description

During bumping the demos to the 22.09  i noticed that the following spec from the Trino tests and docs is invalid
```
apiVersion: trino.stackable.tech/v1alpha1
kind: TrinoCluster
metadata:
  name: simple-trino
spec:
  version: 387-stackable0.1.0
  catalogLabelSelector:
    trino: simple-trino
```
instead it should be 
```
# ...
  catalogLabelSelector:
    matchLabels:
      trino: simple-trino
```
Nobody noticed this, as the field `catalogLabelSelector` is optional. If it is not provided, the (nonexistent) Labelselector will match every TrinoCatalog withing the same Namespace and the test passed anyway.
This behavior sounds unsafe/unwanted so i suggest to make it explicit by requiring a Labelselector.

Although this change is technically breaking, the `TrinoCatalog` feature change has been release for a single day now, so impact should be pretty minimal. It would be greatif we could include this Change in 22.09, so that we reduce the number of breaking changes between Platform releases

## Review Checklist

- [x] Code contains useful comments
- [x] CRD change approved (or not applicable)
- [x] (Integration-)Test cases added (or not applicable)
- [x] Documentation added (or not applicable)
- [x] Changelog updated (or not applicable)
- [x] Cargo.toml only contains references to git tags (not specific commits or branches)
- [x] Helm chart can be installed and deployed operator works (or not applicable)


Once the review is done, comment `bors r+` (or `bors merge`) to merge. [Further information](https://bors.tech/documentation/getting-started/#reviewing-pull-requests)