Discard integration test logs if the upstream Vector aggregator is not available (#407)
# Description

Discard integration test logs if the upstream Vector aggregator is not available

The logging tests fail if the Vector aggregator configured in the environment variable `VECTOR_AGGREGATOR` is not available, i.e. not running or under back pressure because OpenSearch is not available. If `VECTOR_AGGREGATOR` is configured then it is used as the sink for the logs. If this sink is not available then the events are not considered as processed in the transformation steps and the test assertion fails. Therefore `buffer.when_full` is set to `drop_newest` which discards the logs if the buffer is full and yet increases the processed events count which makes the tests pass.